* #sit-issue-552:
  #sit-issue-552 後台【W006 分類階層管理】刪除功能-增加防呆
  #sit-issue-552 後台【W006 分類階層管理】刪除功能-增加防呆

+++ app/Http/Controllers/WebCategoryHierarchyController.php
@@ -154,6 +154,13 @@ class WebCategoryHierarchyController extends Controller
                 'code' => 'E101',
             ], 400);
         }
+        $hasAdSlotContent = $this->webCategoryHierarchyService->hasAdSlotContent($id); 
+        if(!$hasAdSlotContent['status']){
+            return response()->json([
+                'message' => $hasAdSlotContent['message'],
+                'code' => 'E102',
+            ], 400);
+        }
 
         $deleteResult = $this->webCategoryHierarchyService->deleteCategory($id);
